+++ editor/editor-core/src/plugins/media/toolbar/index.tsx
@@ -290,6 +290,7 @@ const generateMediaSingleFloatingToolbar = (
         );
       },
     });
+    toolbarButtons.push({ type: 'separator' });
   }
 
   if (allowAdvancedToolBarOptions) {